pyproject.toml
[project]
name = "linkedin-outreach"
version = "0.1.0"
description = "LinkedIn outreach automation with deterministic scoring, scraping, and CRM sync"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "jinja2>=3.1",
    "playwright>=1.40",
    "patchright>=0.4",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "streamlit>=1.30",
    "loguru>=0.7",
    "apify-client>=1.6",
    "sentence-transformers>=3.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "thefuzz>=0.22",
    "rich>=13.0",
    "anthropic>=0.40",
    "alembic>=1.13",
    "python-dotenv>=1.0",
    "python-jobspy>=1.1",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
outreach = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "live: marks tests that require real browser or network (deselected by default)",
]
addopts = "-m 'not live'"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module-level import not at top (CLI pattern uses sys.path before imports)
    "E741",   # ambiguous variable name (single-letter vars in comprehensions)
    "SIM102", # collapsible-if
    "SIM105", # use contextlib.suppress
    "SIM108", # use ternary operator
    "SIM115", # use context manager for open()
    "SIM116", # use dict instead of if-elif
    "SIM117", # use single with statement
    "RUF001", # ambiguous unicode (en-dash in regex patterns is intentional)
    "RUF012", # mutable class variable without ClassVar
    "B904",   # raise from within except
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811"]  # test files may import for side effects or availability checks
"src/cli/system_commands.py" = ["F821"]  # forward references in string annotations

[tool.ruff.lint.isort]
known-first-party = ["src"]
